requirements.txt
openai
pandas
matplotlib
seaborn
python-dotenv
langchain
langchain_openai
jupyter
notebook
ipykernel